Added Artemis and Apollo3 Board Types
Changed RemoteWidget(QWidget) to MainWindow(QMainWindow) to make it easier to add the Board Type menu and status bar.
The QWidget labels and buttons did not need to be self.'globals'. Removed the self. from those.
Added the Board Type menu with checkboxes for Artemis (for Artemis-based boards) and Apollo3 (for the Edge). Added statusTip's for each.
Added SETTINGS_ARTEMIS.
Refresh (update_com_ports) now restores the previous port if it still exists.
Commented out messages.clear() so previous messages are retained.
Update bootloader baud rate is set automatically to 115200 for Artemis boards and 921600 for Apollo3 boards.